+++ modio/game.py
@@ -111,7 +111,7 @@ class Game(ReportMixin, OwnerMixin):
         self.submitter = None
         self.stats = None
         # self.theme = Theme(**attrs.pop("theme"))
-        self.other_urls = {key: value for key, value in attrs.pop("other_urls")}
+        self.other_urls = {value["label"]: value["url"] for value in attrs.pop("other_urls")}
         self.paltform = [Platform(**platform) for platform in attrs.pop("platforms")]
 
         _submitter = attrs.pop("submitted_by", {})
@@ -393,7 +393,7 @@ class Game(ReportMixin, OwnerMixin):
 
         return Message(**message)
 
-    def add_tag_options(self, name, *, tags=None, hidden=False, tag_type="dropdown"):
+    def add_tag_options(self, name, *, tags=None, hidden=False, locked=False, tag_type="dropdown"):
         """Add tags which mods can apply to their profiles. If the tag names already exists,
         settings such as hidden or type will be overwritten to the values provided and all the
         tags will be added to the group.
@@ -404,12 +404,15 @@ class Game(ReportMixin, OwnerMixin):
         -----------
         name : str
             Name of the tag group
-        type : Optional[str]
+        type : Optional[Literal['dropdown', 'checkboxes']]
             Defaults to dropdown
             dropdown : Mods can select only one tag from this group, dropdown menu shown on site profile.
             checkboxes : Mods can select multiple tags from this group, checkboxes shown on site profile.
         hidden : Optional[bool]
             Whether or not this group of tags should be hidden from users and mod devs. Defaults to False
+        locked : Optional[bool]
+            Whether or not mods can assign from this group of tag to themselves. If locked only game admins
+            will be able to assign the tag. Defaults to False.
         tags : Optional[List[str]]
             Array of tags that mod creators can apply to their mod
         """
@@ -417,7 +420,13 @@ class Game(ReportMixin, OwnerMixin):
             tags = []
 
         tags = {f"tags[{index}]": tag for index, tag in enumerate(tags)}
-        tags = {"name": name, "type": tag_type, "hidden": json.dumps(hidden), **tags}
+        tags = {
+            "name": name,
+            "type": tag_type,
+            "hidden": json.dumps(hidden),
+            "locked": json.dumps(locked),
+            **tags,
+        }
         message = self.connection.post_request(f"/games/{self.id}/tags", data=tags)
 
         tag_option = find(self.tag_options, name=name)
@@ -430,12 +439,20 @@ class Game(ReportMixin, OwnerMixin):
 
         return Message(**message)
 
-    async def async_add_tag_options(self, name, *, tags=None, hidden=False, tag_type="dropdown"):
+    async def async_add_tag_options(
+        self, name, *, tags=None, hidden=False, locked=False, tag_type="dropdown"
+    ):
         if tags is None:
             tags = []
 
         tags = {f"tags[{index}]": tag for index, tag in enumerate(tags)}
-        tags = {"name": name, "type": tag_type, "hidden": json.dumps(hidden), **tags}
+        tags = {
+            "name": name,
+            "type": tag_type,
+            "hidden": json.dumps(hidden),
+            "locked": json.dumps(locked),
+            **tags,
+        }
         message = await self.connection.async_post_request(f"/games/{self.id}/tags", data=tags)
 
         tag_option = find(self.tag_options, name=name)