+++ docs/source/changelog.rst
@@ -27,6 +27,8 @@ New Features
 * `Game.get_stats` renamed to `Game.get_mods_stats`
 * New function `Game.get_stats` that gets stats for the game rather than for the mods of the game
 * New example `examples/polling_events` showing how to use the filter class to only get the latest attributes
+* `Game.add_tag_options` now supports the `locked` option
+* New attribute for TagOption: `locked`
 
 Removed Features
 ##################